Harden deployment core
