+++ internal/platform/db/queries.sql
@@ -117,7 +117,7 @@ ON CONFLICT(name) DO UPDATE SET permissions_json = excluded.permissions_json;
 SELECT * FROM roles WHERE name = ?;
 
 -- name: InsertRoleBinding :exec
-INSERT INTO role_bindings (id, workspace_id, role_id, principal_id, group_id)
+INSERT OR IGNORE INTO role_bindings (id, workspace_id, role_id, principal_id, group_id)
 VALUES (?, ?, ?, ?, ?);
 
 -- name: ListPrincipalRolePermissions :many
