+++ internal/platform/store.go
@@ -178,139 +178,6 @@ func (s *Store) EnsureWorkspace(ctx context.Context, input WorkspaceInput) error
 	})
 }
 
-type PrincipalInput struct {
-	ID          string
-	Email       string
-	DisplayName string
-}
-
-func (s *Store) UpsertPrincipal(ctx context.Context, input PrincipalInput) (db.Principal, error) {
-	if input.ID == "" {
-		input.ID = newID("principal")
-	}
-	if err := s.q.UpsertPrincipal(ctx, db.UpsertPrincipalParams{
-		ID:          input.ID,
-		Email:       input.Email,
-		DisplayName: input.DisplayName,
-	}); err != nil {
-		return db.Principal{}, err
-	}
-	return s.q.GetPrincipal(ctx, input.ID)
-}
-
-func (s *Store) BindRole(ctx context.Context, workspaceID, principalID, roleName string) error {
-	role, err := s.q.GetRoleByName(ctx, roleName)
-	if err != nil {
-		return err
-	}
-	return s.q.InsertRoleBinding(ctx, db.InsertRoleBindingParams{
-		ID:          newID("rolebinding"),
-		WorkspaceID: workspaceID,
-		RoleID:      role.ID,
-		PrincipalID: sql.NullString{String: principalID, Valid: principalID != ""},
-	})
-}
-
-func (s *Store) BootstrapAdmin(ctx context.Context, workspaceID, email string) error {
-	email = strings.TrimSpace(email)
-	if email == "" {
-		return nil
-	}
-	principalID := PrincipalIDForEmail(email)
-	principal, err := s.UpsertPrincipal(ctx, PrincipalInput{ID: principalID, Email: email, DisplayName: email})
-	if err != nil {
-		return err
-	}
-	return s.BindRole(ctx, workspaceID, principal.ID, "owner")
-}
-
-type ExternalIdentityInput struct {
-	Provider    string
-	TenantID    string
-	Subject     string
-	Email       string
-	DisplayName string
-}
-
-func (s *Store) ResolveExternalPrincipal(ctx context.Context, input ExternalIdentityInput) (db.Principal, error) {
-	input.Email = normalizeEmail(input.Email)
-	if input.Provider == "" || input.Subject == "" {
-		return db.Principal{}, fmt.Errorf("external identity requires provider and subject")
-	}
-	identity, err := s.q.GetExternalIdentity(ctx, db.GetExternalIdentityParams{
-		Provider: input.Provider,
-		TenantID: input.TenantID,
-		Subject:  input.Subject,
-	})
-	if err == nil {
-		principal, err := s.UpsertPrincipal(ctx, PrincipalInput{ID: identity.PrincipalID, Email: input.Email, DisplayName: input.DisplayName})
-		if err != nil {
-			return db.Principal{}, err
-		}
-		return principal, nil
-	}
-	if !errors.Is(err, sql.ErrNoRows) {
-		return db.Principal{}, err
-	}
-
-	var principal db.Principal
-	if input.Email != "" {
-		principal, err = s.q.GetPrincipalByEmail(ctx, input.Email)
-		if err != nil && !errors.Is(err, sql.ErrNoRows) {
-			return db.Principal{}, err
-		}
-	}
-	if principal.ID == "" {
-		id := "external_" + stableID(input.Provider+"|"+input.TenantID+"|"+input.Subject)
-		principal, err = s.UpsertPrincipal(ctx, PrincipalInput{ID: id, Email: input.Email, DisplayName: input.DisplayName})
-		if err != nil {
-			return db.Principal{}, err
-		}
-	} else {
-		principal, err = s.UpsertPrincipal(ctx, PrincipalInput{ID: principal.ID, Email: input.Email, DisplayName: input.DisplayName})
-		if err != nil {
-			return db.Principal{}, err
-		}
-	}
-
-	if err := s.q.UpsertExternalIdentity(ctx, db.UpsertExternalIdentityParams{
-		ID:          "identity_" + stableID(input.Provider+"|"+input.TenantID+"|"+input.Subject),
-		PrincipalID: principal.ID,
-		Provider:    input.Provider,
-		TenantID:    input.TenantID,
-		Subject:     input.Subject,
-		Email:       input.Email,
-	}); err != nil {
-		return db.Principal{}, err
-	}
-	return principal, nil
-}
-
-func (s *Store) HasPermission(ctx context.Context, workspaceID, principalID, permission string) (bool, error) {
-	if principalID == "" {
-		return false, nil
-	}
-	rows, err := s.q.ListPrincipalRolePermissions(ctx, db.ListPrincipalRolePermissionsParams{
-		WorkspaceID: workspaceID,
-		PrincipalID: sql.NullString{String: principalID, Valid: true},
-	})
-	if err != nil {
-		return false, err
-	}
-	for _, row := range rows {
-		var permissions []string
-		if err := json.Unmarshal([]byte(row), &permissions); err != nil {
-			return false, err
-		}
-		for _, candidate := range permissions {
-			if candidate == permission {
-				return true, nil
-			}
-		}
-	}
-	return false, nil
-}
-
 func (s *Store) CreateDeployment(ctx context.Context, workspaceID, createdBy string) (db.Deployment, error) {
 	id := newID("dep")
 	if err := s.q.CreateDeployment(ctx, db.CreateDeploymentParams{
@@ -431,108 +298,11 @@ func (s *Store) ActiveArtifact(ctx context.Context, workspaceID string) (db.Depl
 	return deployment, artifact, nil
 }
 
-func (s *Store) CreateSession(ctx context.Context, principalID string, ttl time.Duration) (string, error) {
-	token := newSecret()
-	hash := tokenHash(token)
-	expires := time.Now().Add(ttl).UTC().Format(time.RFC3339)
-	return token, s.q.CreateSession(ctx, db.CreateSessionParams{
-		ID:          newID("session"),
-		PrincipalID: principalID,
-		TokenHash:   hash,
-		ExpiresAt:   expires,
-	})
-}
-
-func (s *Store) PrincipalForToken(ctx context.Context, token string) (db.Principal, error) {
-	session, err := s.q.GetSessionByTokenHash(ctx, tokenHash(token))
-	if err != nil {
-		return db.Principal{}, err
-	}
-	_ = s.q.TouchSession(ctx, session.ID)
-	return s.q.GetPrincipal(ctx, session.PrincipalID)
-}
-
-func (s *Store) DeleteSession(ctx context.Context, token string) error {
-	return s.q.DeleteSessionByTokenHash(ctx, tokenHash(token))
-}
-
-func (s *Store) CreateAPIToken(ctx context.Context, principalID, name string) (string, error) {
-	token := newSecret()
-	return token, s.q.CreateAPIToken(ctx, db.CreateAPITokenParams{
-		ID:          newID("token"),
-		PrincipalID: principalID,
-		Name:        name,
-		TokenHash:   tokenHash(token),
-	})
-}
-
-func (s *Store) PrincipalForAPIToken(ctx context.Context, token string) (db.Principal, error) {
-	apiToken, err := s.q.GetAPITokenByHash(ctx, tokenHash(token))
-	if err != nil {
-		return db.Principal{}, err
-	}
-	_ = s.q.TouchAPIToken(ctx, apiToken.ID)
-	return s.q.GetPrincipal(ctx, apiToken.PrincipalID)
-}
-
 func IgnoreNotFound(err error) error {
 	if errors.Is(err, sql.ErrNoRows) {
 		return nil
 	}
 	return err
-}
-
-type Asset struct {
-	ID           string
-	WorkspaceID  string
-	DeploymentID string
-	Type         string
-	Key          string
-	ParentID     string
-	Title        string
-	Description  string
-	ContentJSON  string
-	ContentHash  string
-}
-
-type AssetEdge struct {
-	ID           string
-	WorkspaceID  string
-	DeploymentID string
-	FromAssetID  string
-	ToAssetID    string
-	Type         string
-}
-
-func NewAsset(workspaceID, deploymentID, typ, key, parentID, title, description string, content any) (Asset, error) {
-	bytes, err := json.Marshal(content)
-	if err != nil {
-		return Asset{}, err
-	}
-	sum := sha256.Sum256(bytes)
-	return Asset{
-		ID:           "asset_" + stableID(deploymentID+"|"+typ+"|"+key),
-		WorkspaceID:  workspaceID,
-		DeploymentID: deploymentID,
-		Type:         typ,
-		Key:          key,
-		ParentID:     parentID,
-		Title:        title,
-		Description:  description,
-		ContentJSON:  string(bytes),
-		ContentHash:  hex.EncodeToString(sum[:]),
-	}, nil
-}
-
-func NewAssetEdge(workspaceID, deploymentID, fromID, toID, typ string) AssetEdge {
-	return AssetEdge{
-		ID:           "edge_" + stableID(deploymentID+"|"+fromID+"|"+toID+"|"+typ),
-		WorkspaceID:  workspaceID,
-		DeploymentID: deploymentID,
-		FromAssetID:  fromID,
-		ToAssetID:    toID,
-		Type:         typ,
-	}
 }
 
 func newID(prefix string) string {
